Fixes method signature for Utils
Previously two disperate objects could be passed into utils, a `Gateway`
or a `provider`.  This changes the signature to always expect a `gateway`
It reduces the potential impact of this change from causing a double
lookup by lazy loading utils.

There are most likely additional refactors that are required here but
this allows the utils class not swallow unrelated errors.